safari-600.7-branch/LayoutTests/accessibility/table-column-headers-with-captions-expected.txt

caption
header1	header2
a	b
This tests that a table with a caption and an empty colgroup will still correctly report its headers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colHeaders[0].isEqual(table.rowAtIndex(0).childAtIndex(0)) is true
PASS colHeaders[1].isEqual(table.rowAtIndex(0).childAtIndex(1)) is true
PASS successfullyParsed is true

TEST COMPLETE
